requirements.txt
Django==1.5.8
South==1.0
django-mongodb-engine==0.5.2
django-redis==3.6.2
django-redis-cache==0.13.0
django-registration==1.0
djangorestframework==2.3.14
djangotoolbox==1.6.2
feedparser==5.1.3
mongoengine==0.8.0
pymongo==2.7.1
redis==2.10.1
wsgiref==0.1.2
